better in progress indicators
